Creating class for firestore services
