+++ lib/utils/services/firestore_services_exceptions.dart
@@ -1,0 +1,6 @@
+class NonExistingDocument implements Exception {
+  @override
+  String toString() {
+    return "Document does not exist";
+  }
+}
